pyproject.toml
[tool.poetry]
name = "aws-cdk-stacks"
version = "0.1.0"
description = "A collection of AWS CDK stacks"
authors = ["garrett-does-code <garrett.codes@gmail.com>"]
readme = "README.md"
packages = [{include = "aws_cdk_stacks"}]

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.92.0"
constructs = "^10.0.0"
pytest = "6.2.5"
python-dotenv = "^1.0.0"
boto3 = "^1.28.31"
pydantic = "^2.4.2"
fastapi = "^0.104.0"
mangum = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    "."
]